chore: relocate config
